% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{batch_get_traces}
\alias{batch_get_traces}
\title{Retrieves a list of traces specified by ID}
\usage{
batch_get_traces(
  TraceIds = list(
    "string"
  ),
  NextToken = "string"
)
}
\arguments{
\item{TraceIds}{[required] Specify the trace IDs of requests for which to retrieve segments.}

\item{NextToken}{Pagination token. Not used.}
}
\description{
Retrieves a list of traces specified by ID. Each trace is a collection of segment documents that originates from a single request. Use \code{GetTraceSummaries} to get a list of trace IDs.
}
\examples{

}
