% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{create_group}
\alias{create_group}
\title{Creates a group resource with a name and a filter expression}
\usage{
create_group(GroupName, FilterExpression = NULL)
}
\arguments{
\item{GroupName}{[required] The case-sensitive name of the new group. Default is a reserved name and names must be unique.}

\item{FilterExpression}{The filter expression defining criteria by which to group traces.}
}
\description{
Creates a group resource with a name and a filter expression.
}
\section{Accepted Parameters}{
\preformatted{create_group(
  GroupName = "string",
  FilterExpression = "string"
)
}
}

