% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{create_sampling_rule}
\alias{create_sampling_rule}
\title{Creates a rule to control sampling behavior for instrumented applications}
\usage{
create_sampling_rule(SamplingRule)
}
\arguments{
\item{SamplingRule}{The rule definition.}
}
\description{
Creates a rule to control sampling behavior for instrumented applications. Services retrieve rules with GetSamplingRules, and evaluate each rule in ascending order of \emph{priority} for each request. If a rule matches, the service records a trace, borrowing it from the reservoir size. After 10 seconds, the service reports back to X-Ray with GetSamplingTargets to get updated versions of each in-use rule. The updated rule contains a trace quota that the service can use instead of borrowing from the reservoir.
}
\examples{

}
