% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{delete_group}
\alias{delete_group}
\title{Deletes a group resource}
\usage{
delete_group(GroupName = NULL, GroupARN = NULL)
}
\arguments{
\item{GroupName}{The case-sensitive name of the group.}

\item{GroupARN}{The ARN of the group that was generated on creation.}
}
\description{
Deletes a group resource.
}
\section{Accepted Parameters}{
\preformatted{delete_group(
  GroupName = "string",
  GroupARN = "string"
)
}
}

