% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{delete_sampling_rule}
\alias{delete_sampling_rule}
\title{Deletes a sampling rule}
\usage{
delete_sampling_rule(RuleName = NULL, RuleARN = NULL)
}
\arguments{
\item{RuleName}{The name of the sampling rule. Specify a rule by either name or ARN, but not both.}

\item{RuleARN}{The ARN of the sampling rule. Specify a rule by either name or ARN, but not both.}
}
\description{
Deletes a sampling rule.
}
\section{Accepted Parameters}{
\preformatted{delete_sampling_rule(
  RuleName = "string",
  RuleARN = "string"
)
}
}

