% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{get_encryption_config}
\alias{get_encryption_config}
\title{Retrieves the current encryption configuration for X-Ray data}
\usage{
get_encryption_config()
}
\description{
Retrieves the current encryption configuration for X-Ray data.
}
\examples{

}
