% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{get_group}
\alias{get_group}
\title{Retrieves group resource details}
\usage{
get_group(GroupName = NULL, GroupARN = NULL)
}
\arguments{
\item{GroupName}{The case-sensitive name of the group.}

\item{GroupARN}{The ARN of the group that was generated on creation.}
}
\description{
Retrieves group resource details.
}
\section{Accepted Parameters}{
\preformatted{get_group(
  GroupName = "string",
  GroupARN = "string"
)
}
}

