% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{get_groups}
\alias{get_groups}
\title{Retrieves all active group details}
\usage{
get_groups(NextToken = NULL)
}
\arguments{
\item{NextToken}{Pagination token. Not used.}
}
\description{
Retrieves all active group details.
}
\section{Accepted Parameters}{
\preformatted{get_groups(
  NextToken = "string"
)
}
}

