% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{get_sampling_rules}
\alias{get_sampling_rules}
\title{Retrieves all sampling rules}
\usage{
get_sampling_rules(NextToken = NULL)
}
\arguments{
\item{NextToken}{Pagination token. Not used.}
}
\description{
Retrieves all sampling rules.
}
\examples{

}
