% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{get_sampling_statistic_summaries}
\alias{get_sampling_statistic_summaries}
\title{Retrieves information about recent sampling results for all sampling rules}
\usage{
get_sampling_statistic_summaries(NextToken = NULL)
}
\arguments{
\item{NextToken}{Pagination token. Not used.}
}
\description{
Retrieves information about recent sampling results for all sampling rules.
}
\section{Accepted Parameters}{
\preformatted{get_sampling_statistic_summaries(
  NextToken = "string"
)
}
}

