% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{get_service_graph}
\alias{get_service_graph}
\title{Retrieves a document that describes services that process incoming requests, and downstream services that they call as a result}
\usage{
get_service_graph(StartTime, EndTime, NextToken = NULL)
}
\arguments{
\item{StartTime}{[required] The start of the time frame for which to generate a graph.}

\item{EndTime}{[required] The end of the time frame for which to generate a graph.}

\item{NextToken}{Pagination token. Not used.}
}
\description{
Retrieves a document that describes services that process incoming requests, and downstream services that they call as a result. Root services process incoming requests and make calls to downstream services. Root services are applications that use the AWS X-Ray SDK. Downstream services can be other applications, AWS resources, HTTP web APIs, or SQL databases.
}
\section{Accepted Parameters}{
\preformatted{get_service_graph(
  StartTime = as.POSIXct("2015-01-01"),
  EndTime = as.POSIXct("2015-01-01"),
  NextToken = "string"
)
}
}

