% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{get_trace_graph}
\alias{get_trace_graph}
\title{Retrieves a service graph for one or more specific trace IDs}
\usage{
get_trace_graph(TraceIds, NextToken = NULL)
}
\arguments{
\item{TraceIds}{Trace IDs of requests for which to generate a service graph.}

\item{NextToken}{Pagination token. Not used.}
}
\description{
Retrieves a service graph for one or more specific trace IDs.
}
\examples{

}
