% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_package.R
\docType{package}
\name{paws.xray-package}
\alias{paws.xray}
\alias{paws.xray-package}
\title{paws.xray: AWS X-Ray}
\description{
AWS X-Ray provides APIs for managing debug traces
and retrieving service maps and other data created by processing those
traces.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
