% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{put_telemetry_records}
\alias{put_telemetry_records}
\title{Used by the AWS X-Ray daemon to upload telemetry}
\usage{
put_telemetry_records(TelemetryRecords, EC2InstanceId = NULL,
  Hostname = NULL, ResourceARN = NULL)
}
\arguments{
\item{TelemetryRecords}{}

\item{EC2InstanceId}{}

\item{Hostname}{}

\item{ResourceARN}{}
}
\description{
Used by the AWS X-Ray daemon to upload telemetry.
}
\examples{

}
