% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{update_group}
\alias{update_group}
\title{Updates a group resource}
\usage{
update_group(GroupName = NULL, GroupARN = NULL,
  FilterExpression = NULL)
}
\arguments{
\item{GroupName}{The case-sensitive name of the group.}

\item{GroupARN}{The ARN that was generated upon creation.}

\item{FilterExpression}{The updated filter expression defining criteria by which to group traces.}
}
\description{
Updates a group resource.
}
\section{Accepted Parameters}{
\preformatted{update_group(
  GroupName = "string",
  GroupARN = "string",
  FilterExpression = "string"
)
}
}

