% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{update_sampling_rule}
\alias{update_sampling_rule}
\title{Modifies a sampling rule's configuration}
\usage{
update_sampling_rule(SamplingRuleUpdate)
}
\arguments{
\item{SamplingRuleUpdate}{The rule and fields to change.}
}
\description{
Modifies a sampling rule's configuration.
}
\examples{

}
