% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{get_sampling_targets}
\alias{get_sampling_targets}
\title{Requests a sampling quota for rules that the service is using to sample requests}
\usage{
get_sampling_targets(SamplingStatisticsDocuments)
}
\arguments{
\item{SamplingStatisticsDocuments}{Information about rules that the service is using to sample requests.}
}
\description{
Requests a sampling quota for rules that the service is using to sample requests.
}
\examples{

}
