% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.xray_operations.R
\name{put_telemetry_records}
\alias{put_telemetry_records}
\title{Used by the AWS X-Ray daemon to upload telemetry}
\usage{
put_telemetry_records(
  TelemetryRecords = list(
    list(
      Timestamp = as.POSIXct("2015-01-01"),
      SegmentsReceivedCount = 123,
      SegmentsSentCount = 123,
      SegmentsSpilloverCount = 123,
      SegmentsRejectedCount = 123,
      BackendConnectionErrors = list(
        TimeoutCount = 123,
        ConnectionRefusedCount = 123,
        HTTPCode4XXCount = 123,
        HTTPCode5XXCount = 123,
        UnknownHostCount = 123,
        OtherCount = 123
      )
    )
  ),
  EC2InstanceId = "string",
  Hostname = "string",
  ResourceARN = "string"
)
}
\arguments{
\item{TelemetryRecords}{[required]}

\item{EC2InstanceId}{}

\item{Hostname}{}

\item{ResourceARN}{}
}
\description{
Used by the AWS X-Ray daemon to upload telemetry.
}
\examples{

}
